Use clang-tidy to auto fix C++ code style issues (this time addressing only superficial problems).

+++ include/swoole_coroutine_c_api.h
@@ -54,7 +54,7 @@ off_t swoole_coroutine_lseek(int fd, off_t offset, int whence);
 int swoole_coroutine_fstat(int fd, struct stat *statbuf);
 int swoole_coroutine_stat(const char *path, struct stat *statbuf);
 int swoole_coroutine_lstat(const char *path, struct stat *statbuf);
-int swoole_coroutine_readlink(const char *pathname, char *buf, size_t len);
+ssize_t swoole_coroutine_readlink(const char *pathname, char *buf, size_t len);
 int swoole_coroutine_unlink(const char *pathname);
 int swoole_coroutine_mkdir(const char *pathname, mode_t mode);
 int swoole_coroutine_rmdir(const char *pathname);
